refactor: 小物関数を string/misc.hpp に集約
lcp/parenthesis（string）と lis/palindrome/rle（algorithm）の単体では
小粒なヘッダを string/misc.hpp に統合し、LCS 接尾辞・split/join・
to_lower/to_upper・next_string などの汎用小物も追加。

- 関数名は既存テストが使う API（run_length_encoding, make_lis,
  is_palindromic_number 等）を踏襲し破壊的変更なし
- 参照していた test 4 本の include を string/misc.hpp に追随
- 元の 5 ファイルを削除（残存参照なしを確認）
